Fix Google browser sign in with Auth0 custom domain (#9)
## The Problem

We updated Auth0 to use a custom domain, but the Electron app is still
listening for redirects to the old Auth0 domain, so the app is no longer
redirecting users to their browser for the Google sign in flow.

## The Solution

Add the new custom Auth0 domain to the list of domains that we detect
when determining whether a redirect should be opened in the browser. I
left the old domain in there for now in case anyone is running old web
code.