Fix bug where log out is sent to browser (#11)
## The Problem

In commit 3275e13b5c4fd44998ec446cbf795550b19bf7d6, I fixed a bug where
we were no longer sending the user to the browser during the Google log
in flow. However, I forgot to add `/authorize` to the authentication URL
string that we were matching against to determine whether or not to send
the redirect to the browser, so it was matching both the log in and the
log out flows. This is preventing users from logging out because the
cookies are removed from their browser instead of the Electron app.

## The Solution

Add `/authorize` to the URL matching string so that only log in
redirects are sent to the browser.