Stop showing HQ window when Swivvel starts (#17)
## The Problem

The HQ window is too distracting as a first user experience. There is a
lot going on and a lot to understand. This distracts from the main
feature: the audio room.

## The Solution

Stop showing the HQ window when Swivvel starts. This will also prevent
the HQ window from opening up during the user onboarding flow.