Close windows on lock; re-open on unlock (#19)
## The Problem

In commit e08c30a4a2d50ef604d2845288fae7e2405cf276, we made it so all of
the Electron windows close when a user's computer goes to sleep and then
re-open when the computer wakes.

There were many reasons for this (see the commit description), but one
was that the app would get in a bad state when users started their day
after keeping the app open all night. We assumed this was because the
computer was going to sleep, but after releasing this fix some users
were still getting into the bad state in the morning. It turns out that
these users were locking their screens, not sleeping their computer.

## The Solution

Update the window close/re-open logic to run both on sleep/wake _and_
lock/unlock.