Buffer idle status change events (#22)
## The Problem

The transparent window is responsible for sending idle change events to
the server. However, sometimes the transparent window isn't available or
is unable to contact the server. This is currently causing us to drop
idle status change events, resulting in users appearing to be idle when
they are not.

## The Solution

To avoid losing idle change events, we now put them all in a buffer and
continually attempt to send that buffer to the transparent window. After
the transparent window successfully syncs an event, it informs the
Electron app, which then removes the event from the buffer.