Handle URL load errors more gracefully; improve logging (#25)
## The Problem

Whenever a window fails to load a URL, we currently show an error
message and quit the entire app. We implemented it this way because this
was the easiest way make sure that the app didn't end up in a broken
state that the user couldn't recover from.

However, we have now encountered a number of situations where quitting
the app is particularly disruptive and confusing to users. For example,
now that we destroy all windows when a user locks their screen, it is
common for the HQ window to fail to load the URL on unlock because the
internet connection hasn't been established yet. In this situation, the
user currently sees a scary error message and the app quits.

## The Solution

Add a new `onError` option to the `loadUrl()` function that allows us to
decide exactly how we want errors to be handled for each window.

## Other Changes

- Remove log line in idle time poller when transparent window is
destroyed. We could get in a state where this would print a log line
every second for hours.
- Add logging when transparent-pixel mouse-over detection changes state.
This will help debug an issue that someone on Ubuntu 22.04 encountered
where the transparent window was always capturing mouse events.
- Prevent the transparent window from ever being hidden. It was possible
to accidentally hide the transparent window using a keyboard shortcut.
- Fix a bug where the app would quit if the transparent window failed to
load a URL after it was destroyed

## Testing

- [x] Simulate login OAuth callback URL failing, app shows error message
and quits
- [x] Simulate settings page failing, app warns and destroys the window
- [x] Simulate HQ window failing, app silently destroys the window
- [x] Transparent window is no longer hidden when using "close window"
keyboard shortcut