Reuse fullscreen bounds logic for HQ window positioning (#27)
## The Problem

When the HQ window is opened, we want to position it to take up the full
screen so that users can see as much of the page as possible. We're
currently doing that by setting the size & position to be the same as
the `workArea` value that we get from the Electron `screen` module.

However, the `workArea` value is sometimes incorrect. On Ubuntu 23.04,
this caused the title bar of the HQ window to be transparent when it
opened, presumably because we were settings the window to be larger than
the available screen area. We previously worked around this by setting
`frame: false` for Linux, but that made it impossible to drag the window
around or close/minimize it using the window title bar buttons.

## The Solution

In commit b6b6d4f941a9fa82098b884e83fd90af54716249, we created a
`getBounds()` function for the transparent window that works around bugs
with the work area size.

This commit moves the `getBounds()` function to a common `utils/`
directory and renames it to `getFullscreenBounds()`, and then uses it
when determining the size and position of the HQ window.

Reusing the `getFullscreenBounds()` function will also help ensure that
any future OS-specific workarounds will be applied everywhere where we
are opening a fullscreen window.

## Other Changes

Stop the transparent window from getting focus when it opens. It was
confusing when the app launched and your current window lost focus
without it being clear what new window gained focus.