Add a ten minute timeout for closing windows on lock and sleep (#26)
## The Problem
Currently we close out of all windows after we detect a lock or sleep
from the device using Swivvel. This leaves a chance for the user to
forget to join back in the `AudioRoom` we booted them out of.

## The Solution
Add a `setTimeout` for 10 minutes that removes them after the 10 minutes
is up, if they open their screen before the 10 minutes the `timeout`
gets cleared, and the windows don't close.