Hide transparent window in screenshots (#28)
## The Problem

The Swivvel audio room widget is always on top of all windows and can be
minimized but not closed completely. This makes it hard to take
screenshots without the widget in them.

## The Solution

Use the `window.setContentProtection(true)` command to prevent the
transparent window from appearing in screenshots. According to the
[Electron
docs](https://www.electronjs.org/docs/latest/api/browser-window#winsetcontentprotectionenable-macos-windows):

> Prevents the window contents from being captured by other apps.
>
> On macOS it sets the NSWindow's sharingType to NSWindowSharingNone. On
Windows it calls SetWindowDisplayAffinity with WDA_EXCLUDEFROMCAPTURE.
For Windows 10 version 2004 and up the window will be removed from
capture entirely, older Windows versions behave as if WDA_MONITOR is
applied capturing a black window.

## Other Changes

- Add a PR template to ensure consistency in PR titles and descriptions
- Add a GitHub action to enforce PR title length