Fix bug where transparent window not always on top on Linux (#30)
## The Problem

In commit 08a3070ffd4ad143cb823e56fef749a3d89e3f29, we made a change to
stop the transparent window from getting focus when it opens. This broke
the always-on-top behavior on Linux.

## The Solution

Call `window.setAlwaysOnTop(true)` after showing the transparent window
to make sure that it is properly configured to be always on top.