Fix `powerMonitor` timeout bugs and add extra logs (#32)
## The Problem
Brian ran into [This
Bug](https://swivvel.slack.com/archives/C041MFH7VMF/p1697127153690059),
which came from the recent change in `handlePowerMonitorStateChanges` -
Where I set it up to close the windows after 10 minutes.

But if the window closed and we still had the timeout, the
`transparentWindow` never opened. Causing a state where we can't get
that open again
<!--
Describe the product or technical problem you are solving with this PR.
Provide as much business context as possible. This problem description
should
be able to stand on its own without linking out to other resources, but
feel
  free to include links to provide helpful context.
-->

## The Solution
Always open the windows even if there is a `timeout`. Since if the
window already exists it will use that one

## Other Changes
Add extra logging with more information so we can properly debug this if
another error occurs.