Report errors to Sentry; add manual error reporting tray item (#31)
## The Problem

We don't have a lot of visibility into errors that users encounter in
the Electron app. Right now our debugging process requires us to ask the
user to explain the issue and manually send us the console logs and the
Electron log file. This makes it difficult to debug issues.

## The Solution

Add automatic Sentry error reporting, as well as a new "Send Bug Report"
item in the tray menu that will attach the Electron `main.log` file with
the Sentry alert.

## Other Changes

Change `npm run dev` to `electron .` because the development environment
wasn't detecting our `package.json` so it was using the app name
`Electron` instead of `Swivvel`. This was causing log files to be
written to an `Electron-development/` directory.