Stop writing idle poller log line every second (#33)
## The Problem

In commit 8bd2bf8890b06995ec449ee8a7d2fe4b673195a0, we attempted to fix
a bug in the idle poller where we could end up writing a log line every
second for hours. When someone locks their screen or suspends their
computer, we close the transparent window. If there is an idle change
event in the buffer, it'll be stuck there until the transparent window
opens again. Sometimes the computer isn't unlocked/resumed for hours,
causing us to write a log line every second for a very long time,
polluting the log.

However, even after that commit there was still a log line in the idle
poller that was getting written every second.

## The Solution

Stop logging how many events are in the buffer unless the transparent
window is available to receive those events.