Use `mousemove` to detect when mouse over transparent area (#36)
## The Problem

Users were reporting that they tried to take a screenshot of a window on
Mac and the screenshot ended up being a blank screen with the widget in
it.

In commit a375ef619c69497c43557c6a61ca2428d37de8a6, we attempted to fix
this by calling `window.setContentProtection(true)` on the transparent
window. This is supposed to prevent apps from taking screenshots of the
transparent window.

However, after releasing that change, users got a new issue when trying
to take screenshots. Instead of being a blank page, they would get an
error.

It turns out that this error only happens when using the "Capture
Selected Window" screenshot option. This option allows you to select a
specific window for the screenshot. But when users tried this, they
could only select the transparent window, not any windows underneath.
This is why they initially got a screenshot of the widget, and then
later got an error because `window.setContentProtection(true)` would no
allow the user to take a screenshot of the transparent window.

## The Solution

After some trial and error, I _think_ that I narrowed the issue down to
calling `transparentWindow.webContents.capturePage()` in a loop. When I
commented out this line, the screenshot tool no longer allowed the user
to select the transparent window as a screenshot target.

Unfortunately, we can't use the polling strategy without calling
`capturePage()`, so I had to see if there were ways I could work around
this. I tried a bunch of things:

- Detect if the user is taking a screenshot and hide the transparent
window. I couldn't find an event for this though. I experimented with
listening for the screenshot keyboard shortcut, but this suffered from
two issues: 1) we wouldn't know when the screenshot was done, 2) users
could initiate a screenshot without using the keyboard shortcuts.
- Blur the transparent window when the mouse is over a transparent
pixel. This seemed to work sometimes, but not consistently.
- Call `transparentWindow.setFocusable(false)` when the mouse is over a
transparent pixel. This didn't work at all, and prevented the window
from ever becoming focused again.
- Add an
[input-event](https://www.electronjs.org/docs/latest/api/web-contents#event-input-event)
listener in the transparent window web contents and only call
`capturePage()` when the mouse is moving. This seemed to work at first
because the input event listener would stop getting called while the
user was selecting the screenshot window target. However, this solution
was also inconsistent. It was possible to get in a state where the
screenshot window selector would still recognize the transparent window
but not the windows underneath.

Eventually, I returned to the Electron GitHub issue that originally
inspired the poll+screenshot strategy and looked for a complete
alternative to the poll+screenshot strategy. I found a strategy that
seems to work well.

The new strategy uses the `forward` option of `setIgnoreMouseEvents()`
to forward mouse events to the window event when mouse events are being
ignored. The web contents register a `mousemove` listener that informs
the Electron window whether the mouse is over the document (i.e. a
transparent area of the window) or a child element (i.e. over an opaque
area of the window). The Electron window then enables or disables the
mouse pass-through on the transparent window.

This strategy is much better than polling because it doesn't suffer from
performance overhead of rapidly taking screenshots of a 1x1 pixel area,
but unfortunately the `forward` option was not working on Windows or
Linux (even though it is documented as being supported on Windows),
meaning that the `mousemove` event in the web contents would stop
triggering after calling `setIgnoreMouseEvents(true)`. So we're
continuing to use the polling strategy for Windows and Linux.