Open transparent window without focus on unlock/resume (#38)
## The Problem

When a user locks or sleeps their computer, we close all windows after
10 minutes. If they unlock or resume their computer within 10 minutes,
we call `windowService.openTransparentWindow()`.

In this scenario, the transparent window already exists, so we end up
just calling `window.show()`. This results in the transparent window
getting focus. We don't want the transparent window to get focus when it
doesn't need to because it can cause weird corner case bugs, e.g. with
screenshots.

## The Solution

Use `showInactive()` to prevent an existing transparent window from
getting focus when opened.

## Other Changes

Minor refactoring of `handlePowerMonitorStateChanges.ts` to remove code
duplication.