Reduce log spam; check for updates every 10 minutes (#39)
## The Problem

There are two events that can cause a lot of spam in our logs:

1. When there are idle events buffered but they can't be handled by the
transparent window (such as when the user is not logged in)
2. When the auto-updater checks for updates

## The Solution

Remove the log line that says when we are sending idle events to the
transparent window. It's not ideal to lose this visibility, but the
cleaner log is likely more important than having this log line. Also, we
still write a log line every time a new event is added to the buffer,
and whenever an event is successfully handled by the transparent window,
so we aren't losing visibility completely.

For the auto-updater log lines, we made two changes:

- Only poll for updates every 10 minutes instead of every minute. We
don't realistically need to be polling every minute. Increasing the poll
interval to 10 minutes does mean that a user will have to have the app
open for longer before the update is detected, but the app is intended
to be open for a long time so this should be an ok tradeoff.
- Don't log when there is no update available. This is already logged by
the auto-updater library, so we were writing two log lines whenever an
update was not available.