Implement screen sharing (#44)
## The Problem

In our own usage of the audio room widget, we are constantly finding
ourselves wanting to screen share. In order to do so, we have to go into
a Zoom breakout room. In most of these cases, we don't care about
enabling video. It would be way better if we could just screen share
directly from the audio room widget.

## The Solution

Add a new button below the mute button that allows users to share their
screen.

## Other Changes

- Update the `no-warning-comment` ESLint rule to warn on the "todo"
string.
- Change the copy for permission requests. The way that macOS prompts
for permission makes it sound like Swivvel wants to access your
microphone _right now_. I figured the description could make it clear
that we won't actually access the microphone until later when the user
takes a specific action.


![image](https://github.com/swivvel/swivvel-electron/assets/119253005/50cd767c-d56b-41b1-9118-aace38d8009d)

- Refactor `useWindowService()` so that log lines written by windows are
always prefixed with the window ID so that it's easier to tell which
windows wrote which log lines.
- Don't attach log files to Sentry alerts in development so that we
don't use up our quota as quickly
- Fix bug where we would throw a ton of errors inside a 50ms interval
loop when the user was prompted to sign in to Google before landing on
the `/getalink` page