Fix google meet window if user isn't signed in to Google (#43)
## The Problem

If a user opens a google meet, but they aren't properly authenticated
with Google, they get an error joining the meet:


https://github.com/swivvel/swivvel-electron/assets/131678810/0a401fe5-68ac-401c-a4b7-789d7821a49c

Electron windows share cookies, so I am not sure how they could be
authenticated with Swivvel but not with Google. But Brian ran into this
yesterday so it is certainly possible

## The Solution

The reason the user gets an error page (best I can tell) is that Google
thinks the user is a bot. If we remove Swivvel and Electron from the
window's user agent, it is just like a user is joining the Meet in an
incognito window, which works fine.

Because the user now has the option to go through the sign in flow, we
have to make sure the patch to getDisplayMedia is injected each time the
page changes.

## Testing


https://github.com/swivvel/swivvel-electron/assets/131678810/7adeac91-6566-4b42-9204-316f767b9f37