Only register transparent window resize handler once (#52)
## The Problem

I was looking at some logs for a user and I saw a ton of log lines like
this, all printed at almost the same time:

> Display change detected
> All displays: [ ... ]

These log lines are printed from our handlers for the `display-added`,
`display-removed`, and `display-metrics-changed` events on the Electron
`screen` module.

It is unlikely that these events were firing so frequently. Instead, I
think that we accidentally registered multiple handlers for each event.
When we open the transparent window, we register handlers for those
events, but then we never _unregister_ the handlers when the transparent
window is destroyed. So if the transparent window is destroyed and
re-opened multiple times, we'll end up with multiple handlers.

## The Solution

Move the handlers for `display-added`, `display-removed`, and
`display-metrics-changed` to be defined just once when the app starts.

## Other Changes

I moved the code in `showOnAllWorkspaces.ts` into
`openTransparentWindow.ts`. We are already setting a lot of other
configuration options on the `BrowserWindow` in
`openTransparentWindow.ts`, so it was weird to have a single option set
from another function.