Ignore auto-update errors; increase update poll to 30 minutes (#53)
## The Problem

We have been getting a ton of Sentry alerts from calls to
`autoUpdater.checkForUpdatesAndNotify()` that are all related to the
user having network connectivity problems. There is nothing we can do
about these issues, so the alerts are currently causing a lot of
alerting noise.

## The Solution

Catch and ignore errors thrown by calls to
`autoUpdater.checkForUpdatesAndNotify()`.

Since we run `checkForUpdatesAndNotify` synchronously when the app
starts up, I also moved our `configureAutoUpdates()` function in
`background.ts` to be called at the very end so that it won't block any
other setup if it happens to take a while to execute the function.

## Other Changes

I also opted to increase the length of time between checks for new
versions from 10 minutes to 30 minutes. This was a somewhat arbitrary
change not prompted by anything specific, aside from further reducing
the number of network requests that our app does and further reducing
logging noise.