Handle log in from transparent window flow (#55)
## The Problem

When a user first opens the app they aren't logged in, so they are
presented with the login window. They must then go through the login
process _and_ the setup flow before they get a taste of the widget,
which is what Swivvel actually is. The standalone log in window also
makes the auth flow feel a bit disjointed.

## The Solution

If a user is unauthenticated, we'll show the log in button in the
transparent window. When it is clicked, it will fire an event back to
electron which will kick of the log in flow, which will open the login
flow in the browser. This removes the need for the log in window (except
for local dev). See the README in the main repo for a full explanation
of the login flow