Always show Dev Tools tray menu option (#57)
## The Problem

We have been getting some bug reports that have been very difficult to
reproduce. Ideally we would get on a screen share with the user, but we
would need access to the dev tools on the transparent window for the
screen share to be effective. On Windows and Linux there is no easy way
to open the transparent window dev tools. Also, sometimes the issue
we're trying to debug involves the audio room widget disappearing
completely, making it difficult to open dev tools on Mac as well.

## The Solution

Always show the Dev Tools tray menu option.