Add logs to debug issue where `window.electron` is undefined (#59)
## The Problem

We have an Ubuntu 23.04 user who is opening Swivvel in the desktop app,
but the app thinks that it is not running in the desktop app because
`window.electron` is undefined. So one of the following must be
happening

- The preload script is failing to get included in the browser window
options (perhaps there is an issue with the path)
- The preload script is getting included but is not running
- The preload script is running but `contextBridge.exposeInMainWorld()`
is not working

## The Solution

Add logs so we can verify the path of the preload script and make sure
that it is successfully running.