Add keyboard shortcut for toggling mute state (#61)
## The Problem

When someone starts talking to you in Swivvel, it can be annoying to try
to get to the widget and click unmute, especially if the widget is
currently collapsed.

## The Solution

Register `Alt+Shift+U` as a keyboard shortcut for muting and unmuting.

I tried this keyboard shortcut on Linux, Mac and Windows and it does not
seem to be registered on any of them (though individual apps and
websites might use it).

## Other Changes

In commit 5efb90a1263c70cd34312be338f377fc44ca42d0 we upgraded the
`electron-log` package. One of the changes was to deprecate the
`resolvePath` field on the file transport and rename it to
`resolvePathFn`. There is still one place we are using the old
`resolvePath` name, so I updated it to be `resolvePathFn`.