Allow link to google in meet window for Okta sign in (#64)
## The Problem

To use Google Meet in Swivvel, users need to sign in to Google
workspace. Some companies will use okta to sign in. As a part of the
okta sign in flow, the user is redirected to a url on the `google.com`
domain. Because that domain has not been safelisted, electron kicked it
out to the browser which prevents the sign in flow from completing

## The Solution

Allow the url to open in the electron window