Use maximized temp window to get work area bounds on Linux (#68)
## The Problem

After just the first few handful of Ubuntu users (spanning versions from
22.04 to 23.10), it's clear that the work area bounds are not reliable.
In many cases, the work area would fail to take into account the top bar
or the Dock. Sometimes it would be correct for one display but incorrect
for a second display.

At first we tried to guess the bounds by looking at the version of
Ubuntu and accounting for any observed issues with the work area values.
However, there are many different ways each version can be configured
(not to mention the variety of other distros), so it isn't tenable to
create workarounds for every possible configuration.

## The Solution

We attempted to sidestep this problem completely by simply maximizing
the transparent window. However, the window wouldn't maximize because we
set the `type=splash` and for some reason this blocks maximization. We
tried removing the `type` field, but then the `alwaysOnTop` behavior
stopped working. We attempted a variety of other configurations, but we
could not get the transparent window to maximize without introducing
other bugs.

The only working approach we have found is to calculate the work area
bounds by creating a temporary window, maximizing it, then reading the
size and position of that window. This way we don't have to tweak any
config options on the transparent window but we can still use OS-level
window maximization to get the correct bounds for every possible version
and configuration of Linux.

## Other Changes

Stop formatting the browser window options object when logging it. It's
a little easier to scan the logs when there aren't large gaps due to big
formatted objects.