Fix transparent window resize on display change on Linux (#69)
## The Problem

In commit b96ce9d2ab90521a5b8b79321339e62be9e7d77f, we updated the
positioning strategy for the transparent window on Linux. Instead of
relying on the work area bounds (which is unreliable), we now create a
temporary invisible window, maximize it, and then read the bounds.

However, there are now a few issues when the transparent window is
resized due to a display change (e.g. plugging in an external monitor):

1. Sometimes the temporary window doesn't get closed so it just hangs
around
2. Maximizing the temporary window takes a few seconds, during which
time the widget might be floating in the middle of the screen
3. In some cases the widget is rendered off screen after a display
change. This seems to happen when the resolution width goes down.

## The Solution

A couple small tweaks address the issues above:

1. Stop relying on the `maximize` event on the temp window. This wasn't
consistently firing, which is what was causing the temp windows to hang
around sometimes. Instead, just maximize the window and wait two
seconds. This isn't ideal, but hopefully two seconds will be enough for
the window to maximize successfully so that we can read the new bounds.
2. Hide the transparent window while we are calculating the new bounds
so that the widget doesn't appear floating in the middle of the screen
3. Instead of calling `transparentWindow.setSize()` and
`transparentWindow.setPosition()` separately, perform the repositioning
with a single call to `transparentWindow.setBounds()`. Some OS's will
disregard a window resize or reposition if it would cause the window the
exceed the bounds of the screen. This is what was causing the widget to
render off screen because the `setSize()` was being ignored. By calling
`setBounds()`, we can ensure that the window will never be off-screen.