Ignore screen capture errors in mouse pass-through handler (#72)
## The Problem

Sometimes we get the following error: "Current display surface not
available for capture". We're not sure why this happens, but we got this
error once during development and it seemed to be related to the
transparent window being hidden or in a similar state.

## The Solution

If we fail to capture a screenshot, there isn't much we can do about it
so we're ignoring the error to reduce alerting noise and ignoring mouse
events so that the user can interact with the rest of their system.