Re-enable content protection to prevent blue bar on windows (#73)
## The Problem

In commit 7906dc07df272e71ebf24f57b0668eace47bdfe0, we removed the
following setting on the transparent window:

```ts
window.setContentProtection(true);
```

The intention was to make it so that users could take screenshots of the
widget when filing bug reports.

However, there was an unintended consequence. Windows users started
reporting a blue bar at the top of their screen:


![image](https://github.com/swivvel/swivvel-electron/assets/119253005/150476b9-7086-47b8-bc1b-2a69b148005a)

## The Solution

I have no clue why this blue bar appears when content protection is off,
but I'm turning content protection back on for now to fix the bug, and
then we can take some time to figure out why the blue bar is appearing.