Add "Upgrade" option to tray menu (#74)
## The Problem

We frequently have to tell users to upgrade the desktop app. Every
running version of the app will upgrade automatically at midnight, but
sometimes we need users to upgrade before then, such as when we pushed a
bug fix.

Right now, the only way to ensure that a user is on the latest version
is to tell them to restart the app twice (once to download the update, a
second time to install it). This is somewhat awkward.

## The Solution

Add a "Check for Updates" option to the tray menu that checks for a new
update and prompts the user to download it if one is available.


![image](https://github.com/swivvel/swivvel-electron/assets/119253005/382aa85a-363f-4a41-92a7-5a69487b2580)


![image](https://github.com/swivvel/swivvel-electron/assets/119253005/b58709d5-37b3-45ed-96e4-6fad0776821e)


![image](https://github.com/swivvel/swivvel-electron/assets/119253005/2fcdec1d-2102-4c86-9b34-0585bde40903)