Log OS and hardware specs when app starts (#76)
## The Problem

Some users have mentioned that the app is causing their computer to run
slowly. It would be handy to have an easy way to get information about
each user's OS and hardware to help identify a possible pattern and try
to reproduce the issue.

## The Solution

Log some information about the user's system when the app starts.