Prevent widget from flickering on minor display changes (#78)
## The Problem

I was looking at some logs and noticed the following series of log lines

```
[2023-11-21 11:54:29.045] [info]  [transparent] All displays: [{"id":1,"label":"Built-in Retina Display","bounds":{"x":0,"y":0,"width":1470,"height":956},"workArea":{"x":0,"y":37,"width":1470,"height":863},"accelerometerSupport":"unknown","monochrome":false,"colorDepth":30,"colorSpace":"{primaries:BT709, transfer:SRGB_HDR, matrix:RGB, range:FULL}","depthPerComponent":10,"size":{"width":1470,"height":956},"displayFrequency":60,"workAreaSize":{"width":1470,"height":863},"scaleFactor":2,"rotation":0,"internal":true,"touchSupport":"unknown"}]
...
[2023-11-21 11:54:34.564] [info]  Display change detected
[2023-11-21 11:54:34.564] [info]  All displays: [{"id":1,"label":"Built-in Retina Display","bounds":{"x":0,"y":0,"width":1470,"height":956},"workArea":{"x":0,"y":37,"width":1470,"height":864},"accelerometerSupport":"unknown","monochrome":false,"colorDepth":30,"colorSpace":"{primaries:BT709, transfer:SRGB_HDR, matrix:RGB, range:FULL}","depthPerComponent":10,"size":{"width":1470,"height":956},"displayFrequency":60,"workAreaSize":{"width":1470,"height":864},"scaleFactor":2,"rotation":0,"internal":true,"touchSupport":"unknown"}]
...
[2023-11-21 11:54:42.914] [info]  Display change detected
[2023-11-21 11:54:42.916] [info]  All displays: [{"id":1,"label":"Built-in Retina Display","bounds":{"x":0,"y":0,"width":1470,"height":956},"workArea":{"x":0,"y":37,"width":1470,"height":865},"accelerometerSupport":"unknown","monochrome":false,"colorDepth":30,"colorSpace":"{primaries:BT709, transfer:SRGB_HDR, matrix:RGB, range:FULL}","depthPerComponent":10,"size":{"width":1470,"height":956},"displayFrequency":60,"workAreaSize":{"width":1470,"height":865},"scaleFactor":2,"rotation":0,"internal":true,"touchSupport":"unknown"}]
...
```

Notice that there were no displays added or removed, and the only change
to the single display appears to be the work area size changing by 1
pixel.

In commit b47000db1c1a8aaac0c4939c7cbff20e97fa4052, we updated the
transparent window resize handler to hide the window after a display
change was detected, and then re-show the window after the new bounds
have been calculated. This was to prevent the widget from floating in
the middle of the screen after a display was added or removed and before
the new bounds had been computed.

However, we didn't realize that the `display-metrics-changed` will
sometimes get triggered rapidly for tiny changes to the work area size.
This likely made it so the user saw the widget flickering in and out
every time a display change was detected.

## The Solution

Only hide the transparent window if the display change is something
major, i.e. a display being added or removed.

I also added the display change event name to the logs so that we have
more information about what triggered the transparent window resize.